firmware: Restructure timera code a bit
This makes the action to be taken by the timera interrupt handler
explicit in a new global variable. This makes the code a bit more clear.

Also, currently the shift to the next bit only happens when the timera
interrupt fires, even though it could sometimes happen in the int0
interrupt already. Additionally, in some cases the timera interrupt
doesn't even need to fire (saving power). This change prepares for
making that possible.

On the downside, this commit makes the code a bit bigger (but only 18
bytes, part of which is because GCC is missing some optimizations...).