firmware: Replace FLAG_RECEIVE with FLAG_IDLE
Previously, the idle state was entered when neither FLAG_SEND or
FLAG_RECEIVE was set. However, it's not common to enter idle mode (and
if it happens, the value of flags is no longer relevant). So, giving it
an explicit bit that can be assumed to be normally not set makes the
code a bit more compact (previously, all code had to either set
FLAG_SEND and reset FLAG_RECEIVE or vice versa, now it only needs to set
or reset FLAG_SEND).